fix metrics api
