+++ framework/src/main/java/org/tron/core/metrics/MetricsConfig.java
@@ -1,11 +1,11 @@
-package org.tron.core.services.monitor;
+package org.tron.core.metrics;
 
 import com.codahale.metrics.MetricRegistry;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class MonitorConfig {
+public class MetricsConfig {
   @Bean
   public MetricRegistry metricRegistry() {
     return new MetricRegistry();
