+++ framework/src/main/java/org/tron/core/metrics/MonitorMetric.java
@@ -1,16 +1,18 @@
-package org.tron.core.services.monitor;
+package org.tron.core.metrics;
 
-import com.codahale.metrics.Histogram;
-import com.codahale.metrics.Meter;
-import com.codahale.metrics.MetricRegistry;
+import com.codahale.metrics.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+
+import java.util.SortedMap;
 
 @Component
 public class MonitorMetric {
 
   public static final String NET_BLOCK_LATENCY = "net.block.latency";
   public static final String BLOCKCHAIN_TPS = "blockchain.TPS";
+  public static final String NET_DISCONNECTION_COUNT = "net.disconnection.count";
+  public static final String NET_DISCONNECTION_REASON = "net.disconnection.reason";
 
   @Autowired
   private MetricRegistry metricRegistry;
@@ -21,5 +23,13 @@ public class MonitorMetric {
 
   public Meter getMeter(String name) {
     return metricRegistry.meter(name);
+  }
+
+  public Counter getCounter(String name) {
+    return metricRegistry.counter(name);
+  }
+
+  public SortedMap<String, Counter> getCounters(String name) {
+    return metricRegistry.getCounters((s, metric) -> s.startsWith(name));
   }
 }
