+++ framework/src/main/java/org/tron/core/metrics/NetInfo.java
@@ -1,0 +1,5 @@
+package org.tron.core.metrics;
+
+public class NetInfo {
+
+}
