+++ framework/src/main/java/org/tron/core/net/TronNetDelegate.java
@@ -48,7 +48,7 @@ import org.tron.core.net.message.BlockMessage;
 import org.tron.core.net.message.MessageTypes;
 import org.tron.core.net.message.TransactionMessage;
 import org.tron.core.net.peer.PeerConnection;
-import org.tron.core.services.monitor.MonitorMetric;
+import org.tron.core.metrics.MonitorMetric;
 import org.tron.core.store.WitnessScheduleStore;
 import org.tron.protos.Protocol.Inventory.InventoryType;
