--- framework/src/main/java/org/tron/core/services/monitor/NetInfo.java
@@ -1,5 +1,0 @@
-package org.tron.core.services.monitor;
-
-public class NetInfo {
-
-}
