+++ protocol/src/main/protos/core/Tron.proto
@@ -633,97 +633,97 @@ message NodeInfo {
   }
 }
 
-message MonitorInfo {
-  int32 status = 1;
-  string msg = 2;
-  DataInfo data = 3;
+message MetricsInfo {
+  int32 interval = 1;
+  NodeInfo node = 2;
+  BlockChainInfo blockchain = 3;
+  NetInfo net = 4;
 
-  message DataInfo {
-    int32 interval = 1;
-    NodeInfo node = 2;
-    BlockChainInfo blockchain = 3;
-    NetInfo net = 4;
+  message NodeInfo {
+    string ip = 1;
+    int32 nodeType = 2;
+    int32 status = 3;
+    string version = 4;
+  }
 
-    message NodeInfo {
-      string ip = 1;
-      int32 Type = 2;
-      int32 status = 3;
-      string version = 4;
-      int32 noUpgradedSRCount = 5;
-      repeated NoUpgradedSR noUpgradedSRList = 6;
-      message NoUpgradedSR {
-        string address = 1;
-        string url = 2;
-      }
+  message BlockChainInfo {
+    int64 headBlockNum = 1;
+    int64 headBlockTimestamp = 2;
+    string headBlockHash = 3;
+    int32 forkCount = 4;
+    TPSInfo blockProcessTime = 5;
+    TPSInfo TPS = 6;
+    int32 TxCacheSize = 7;
+    int32 missedTxCount = 8;
+    repeated Witness witnesses = 9;
+    message Witness {
+      string address = 1;
+      int32 version = 2;
     }
-
-    message BlockChainInfo {
-      int64 headBlockNum = 1;
-      int64 headBlockTimestamp = 2;
-      string headBlockHash = 3;
-      int32 forkCount = 4;
-      int32 blockProcessTime = 5;
-      TPSInfo TPS = 6;
-      int32 TxCacheSize = 7;
-      int32 missedTxCount = 8;
-      message TPSInfo {
-        double meanRate = 1;
-        double oneMinuteRate = 2;
-        double fiveMinuteRate = 3;
-        double fifteenMinuteRate = 4;
-      }
-    }
-
-    message NetInfo {
-      int32 errorProtoCount = 1;
-      ApiInfo api = 2;
-      int32 connectionCount = 3;
-      int32 validConnectionCount = 4;
-      int64 TCPInTraffic = 5;
-      int64 TCPOutTraffic = 6;
-      int32 disconnectionCount = 7;
-      repeated DisconnectionDetailInfo disconnectionDetail = 8;
-      int64 UDPInTraffic = 9;
-      int64 UDPOutTraffic = 10;
-      LatencyInfo latency = 11;
-
-      message ApiInfo {
-        int32 totalCount = 1;
-        int32 totalFailCount = 2;
-        repeated ApiDetailInfo detail = 3;
-        message ApiDetailInfo {
-          string name = 1;
-          int32 count = 2;
-          int32 failCount = 3;
-        }
-      }
-
-      message DisconnectionDetailInfo {
-        string reason = 1;
-        int32 count = 2;
-      }
-
-      message LatencyInfo {
-        int32 top99 = 1;
-        int32 top95 = 2;
-        int32 totalCount = 3;
-        int32 delay1S = 4;
-        int32 delay2S = 5;
-        int32 delay3S = 6;
-        repeated LatencyDetailInfo detail = 7;
-
-        message LatencyDetailInfo {
-          string witness = 1;
-          int32 top99 = 2;
-          int32 top95 = 3;
-          int32 count = 4;
-          int32 delay1S = 5;
-          int32 delay2S = 6;
-          int32 delay3S = 7;
-        }
-      }
+    message TPSInfo {
+      double meanRate = 1;
+      double oneMinuteRate = 2;
+      double fiveMinuteRate = 3;
+      double fifteenMinuteRate = 4;
     }
   }
 
+  message NetInfo {
+    int32 errorProtoCount = 1;
+    ApiInfo api = 2;
+    int32 connectionCount = 3;
+    int32 validConnectionCount = 4;
+    RateInfo TCPInTraffic = 5;
+    RateInfo TCPOutTraffic = 6;
+    int32 disconnectionCount = 7;
+    repeated DisconnectionDetailInfo disconnectionDetail = 8;
+    RateInfo UDPInTraffic = 9;
+    RateInfo UDPOutTraffic = 10;
+    LatencyInfo latency = 11;
+
+    message ApiInfo {
+      int32 totalCount = 1;
+      int32 totalFailCount = 2;
+      repeated ApiDetailInfo detail = 3;
+      message ApiDetailInfo {
+        string name = 1;
+        int32 count = 2;
+        int32 failCount = 3;
+      }
+    }
+
+    message DisconnectionDetailInfo {
+      string reason = 1;
+      int32 count = 2;
+    }
+
+    message LatencyInfo {
+      int32 top99 = 1;
+      int32 top95 = 2;
+      int32 totalCount = 3;
+      int32 delay1S = 4;
+      int32 delay2S = 5;
+      int32 delay3S = 6;
+      repeated LatencyDetailInfo detail = 7;
+
+      message LatencyDetailInfo {
+        string witness = 1;
+        int32 top99 = 2;
+        int32 top95 = 3;
+        int32 count = 4;
+        int32 delay1S = 5;
+        int32 delay2S = 6;
+        int32 delay3S = 7;
+      }
+    }
+
+
+    message RateInfo {
+      double meanRate = 1;
+      double oneMinuteRate = 2;
+      double fiveMinuteRate = 3;
+      double fifteenMinuteRate = 4;
+    }
+  }
 }
